requirements.txt
blinker==1.8.2
boto3==1.35.50
botocore==1.35.50
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
totoapicontroller==1.1.2
tzdata==2024.2
urllib3==1.26.20
Werkzeug==3.0.6
zipp==3.20.2
